+++ graphqlSchema/core/types/userPayment/UserPayment.js
@@ -14,7 +14,7 @@ const UserPayment = `
   {
     user: User! @relation(name: "UserPayment", direction: "OneWay")
     product: Product! @relation(name: "ProductUserPayment", direction: "OneWay")
-    amount: Int!
+    amount: Float!
     isDiscountUsed: Boolean @defaultValue(value: "false")
     discount: Discount @relation(name: "Discount", direction: "OneWay")
     status: String!
